Extract ProgressDialog so it can be reused
Issue: #239
